Move scheme implementation into scheme.rs, remove unused functions
